Add files via upload
This is going to be something that gonna contain each and every project file paste and extract it as it is on the xampp or wampp folder. And then take the sql file which i uploaded separately and then create a db and then make sure to upload this file. And this project will be up and running